[m]: Adds image compression

+++ discussions/models.py
@@ -1,7 +1,11 @@
 from django.contrib.auth.models import User
 from django.db import models
 from django.utils.text import slugify
-from ckeditor.fields import RichTextField, CKEditorWidget
+from ckeditor.fields import RichTextField
+import sys
+from PIL import Image
+from io import BytesIO
+from django.core.files.uploadedfile import InMemoryUploadedFile
 
 CATEGORY_TYPE = (
     ("Apparels & Accessories", "Apparels & Accessories"),
@@ -22,6 +26,18 @@ CATEGORY_TYPE = (
     ("Toys & Video Games", "Toys & Video Games"),
     ("Travel, Tour & Packages", "Travel, Tour & Packages"),
     ("Others", "Others"))
+
+
+def compress_image(img):
+    size = 1080, 960
+    image_temporary = Image.open(img).convert('RGB')
+    output_io_stream = BytesIO()
+    image_temporary.thumbnail(size, Image.ANTIALIAS)
+    image_temporary.save(output_io_stream, format='JPEG', quality=75)
+    output_io_stream.seek(0)
+    img = InMemoryUploadedFile(output_io_stream, 'ImageField', "%s.jpg" % img.name.split('.')[0], 'image/jpeg',
+                               sys.getsizeof(output_io_stream), None)
+    return img
 
 
 class Discussion(models.Model):
@@ -55,6 +71,8 @@ class Discussion(models.Model):
     def save(self, *args, **kwargs):
         if not self.slug:
             self.slug = self._get_unique_slug()
+        if self.img:
+            self.img = compress_image(self.img)
         super().save(*args, **kwargs)
 
     def get_tags(self):
@@ -62,7 +80,8 @@ class Discussion(models.Model):
         return tag
 
     def delete(self, *args, **kwargs):
-        self.img.delete()
+        if self.img:
+            self.img.delete()
         super().delete(*args, **kwargs)
